revert: drop cross-board saved Views (#3815) from this branch
The Views feature as built (a flat, virtualized cross-board card LIST in the
left nav) doesn't match the intended product: a saved filter that opens as a
full board-like surface with the Kanban/List/Timeline display switcher, listed
in the left nav like boards. Reverting it out of this PR so the full-page card
view (#3817) and personal reminders (#3816) can ship on their own; Views will be
re-scoped and rebuilt on a separate branch.

This reverts commit c05d0bc (feat(views): cross-board saved Views) — removes
ViewService/ViewController/CardSummaryService/ViewsView/useViews + the /views
route, nav entry, API calls, tests, and CHANGELOG entry. BoardController returns
to its exact pre-Views (inline serializeCardSummaries) state; reminders and the
card-view fix are untouched.
