+++ tests/unit/Controller/BoardControllerTest.php
@@ -31,7 +31,6 @@ use OCA\Kanso\Db\ReviewTypeMapper;
 use OCA\Kanso\Db\Stack;
 use OCA\Kanso\Db\StackMapper;
 use OCA\Kanso\Service\BoardService;
-use OCA\Kanso\Service\CardSummaryService;
 use OCA\Kanso\Service\ContactService;
 use OCA\Kanso\Service\InvalidInputException;
 use OCA\Kanso\Service\NotPermittedException;
@@ -105,21 +104,6 @@ class BoardControllerTest extends TestCase {
 		$user->method('getUID')->willReturn('alice');
 		$this->userSession->method('getUser')->willReturn($user);
 
-		// The per-card enrichment now lives in a shared CardSummaryService (#3815),
-		// reused by both the board payload and the cross-board Views feed. Wire the
-		// same enrichment-mapper mocks into a REAL instance so the existing show()/
-		// changes() enrichment assertions below exercise the identical fold.
-		$cardSummaryService = new CardSummaryService(
-			$this->cardLabelMapper,
-			$this->cardAssigneeMapper,
-			$this->cardContactMapper,
-			$this->checklistItemMapper,
-			$this->cardMapper,
-			$this->commentMapper,
-			$this->cardReviewMapper,
-			$this->cardRelationMapper,
-		);
-
 		$this->controller = new BoardController(
 			'kanso',
 			$this->request,
@@ -131,13 +115,19 @@ class BoardControllerTest extends TestCase {
 			$this->stackMapper,
 			$this->cardMapper,
 			$this->labelMapper,
+			$this->cardLabelMapper,
+			$this->cardAssigneeMapper,
+			$this->cardContactMapper,
+			$this->cardReviewMapper,
 			$this->reviewTypeMapper,
 			$this->cardFieldMapper,
+			$this->checklistItemMapper,
+			$this->commentMapper,
 			$this->aclMapper,
 			$this->permissionService,
 			$this->subscriptionService,
-			$this->boardAccess,
-			$cardSummaryService
+			$this->cardRelationMapper,
+			$this->boardAccess
 		);
 	}
